remove utility functions & bs58 dep
These are all just functions for converting between representations of a
multihash which is absurd to include because js-multihash only needs to
do conversion between a Buffer and a parsed object. There's no reason to
bundle one specific library for encoding/decoding base-58. Since base-58
encoding is a well-defined operation, users should be able to choose
whatever base-58 library they want.

The functions `toHexString` & `fromHexString` are even worse because
they just add a pointless wrapper to the builtin `Buffer` functions and
make any code that uses them harder to understand.